Add theme utility module
